Added lots of test cases and update readme.
