+++ README.md
@@ -74,23 +74,23 @@ Currently, there is a minimal subset of the NEO protocol implemented.
 To start experimenting make sure you a have a private net running on your machine. 
 If you dont, take a look at [docker-privnet-with-gas](https://hub.docker.com/r/metachris/neo-privnet-with-gas/). 
 
-Start the server:
+Start a NEO node:
 
 ```
-./bin/neo-go -seed 127.0.0.1:20333
+./bin/neo-go node -seed 127.0.0.1:20333
 ```
 
 You can add multiple seeds if you want:
 
 ```
-./bin/neo-go -seed 127.0.0.1:20333,127.0.01:20334
+./bin/neo-go node -seed 127.0.0.1:20333,127.0.01:20334
 ```
 
 By default the server will currently run on port 3000, for testing purposes. 
 You can change that by setting the tcp flag:
 
 ```
-./bin/neo-go -seed 127.0.0.1:20333 -tcp 1337
+./bin/neo-go node -seed 127.0.0.1:20333 -tcp 1337
 ```
 
 ## RPC
@@ -98,10 +98,10 @@ You can change that by setting the tcp flag:
 If you want your node to also serve JSON-RPC, you can do that by setting the following flag:
 
 ```
-./bin/neo-go -rpc 4000
+./bin/neo-go node -rpc 4000
 ```
 
-In this case server will accept and respond JSON-RPC on port 4000. 
+In this case the server will accept and respond JSON-RPC on port 4000. 
 Keep in mind that currently there is only a small subset of the JSON-RPC implemented. 
 Feel free to make a PR with more functionality.
 
@@ -112,9 +112,42 @@ TODO
 ```
 
 ## Smart Contracts
+> NOTE: At this moment there is only a small subset of the Go language implemented. 
+
+### Compile a smart contract
 
 ```
-TODO
+./bin/neo-go contract compile mycontract.go
+```
+
+By default the filename will be the name of your .go file with the .avm extension, the file will be located in the same directory where you called the command from. If you want another location for your compiled contract:
+
+```
+./bin/neo-go contract compile mycontract.go --out /Users/foo/bar/contract.avm
+```
+
+### Debugging your smart contract
+You can dump the opcodes generated by the compiler with the following command:
+
+```
+./bin/neo-go contract opdump mycontract.go
+```
+
+This will result in something like this:
+
+```
+INDEX    OPCODE    DESC
+0        0x52      OpPush2
+1        0xc5      OpNewArray
+2        0x6b      OpToAltStack
+3        0x 0      OpPush0
+4        0x6c      OpFromAltStack
+5        0x76      OpDup
+6        0x6b      OpToAltStack
+7        0x 0      OpPush0
+8        0x52      OpPush2
+9        0x7a      OpRoll
+10       0xc4      OpSetItem
 ```
 
 # Contributing
