+++ pkg/vm/compiler/tests/string_test.go
@@ -1,0 +1,15 @@
+package compiler_test
+
+var stringTestCases = []testCase{
+	{
+		"simple string",
+		`
+		package testcase
+		func Main() string {
+			x := "NEO"
+			return x
+		}
+		`,
+		"52c56b034e454f6c766b00527ac46203006c766b00c3616c7566",
+	},
+}
